task 1

+++ 0x14-bit_manipulation/main.h
@@ -2,7 +2,7 @@
 #define MAIN_H_
 
 unsigned int binary_to_uint(const char *b);
-unsigned int _power(unsigned int base, unsigned int pow);
+unsigned long int _power(unsigned int base, unsigned int pow);
 void print_binary(unsigned long int n);
 int _putchar(char c);
